requirements.txt
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
aiosqlite==0.17.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.2
asgiref==3.6.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7